+++ js/main.js
@@ -1,4 +1,4 @@
-import {ANNOUNCEMENT_QUANTITY} from './create-announcement.js';
 import {moreAnnouncement} from './create-announcement.js';
+const ANNOUNCEMENT_QUANTITY = 10;
 
 moreAnnouncement(ANNOUNCEMENT_QUANTITY);
