--- js/random-massive.js
@@ -1,10 +1,0 @@
-// Функция копирует, затем перемешивает массив в случайном порядке
-const getShuffledItems = (items) => items.slice().sort(() => Math.random() - 0.5);
-
-// Функция возвращает перемешенный массив до заданной длины массива
-const getRandomItems = (items, count) => {
-  const shuffledItems = getShuffledItems(items);
-  return shuffledItems.slice(0, count);
-};
-
-export {getRandomItems};
