+++ js/util-function.js
@@ -1,3 +1,12 @@
+// Функция копирует, затем перемешивает массив в случайном порядке
+const getShuffledItems = (items) => items.slice().sort(() => Math.random() - 0.5);
+
+// Функция возвращает перемешенный массив до заданной длины массива
+const getRandomItems = (items, count) => {
+  const shuffledItems = getShuffledItems(items);
+  return shuffledItems.slice(0, count);
+};
+
 const getRandomNumber = (min, max) => {
   const lower = Math.ceil(Math.min(Math.abs(min), Math.abs(max)));
   const upper = Math.floor(Math.max(Math.abs(min), Math.abs(max)));
@@ -12,4 +21,4 @@ const getRandomFractionNumber = (min, max, digits = 1) => {
   return result.toFixed(digits);
 };
 
-export {getRandomNumber, getRandomFractionNumber};
+export {getRandomNumber, getRandomFractionNumber, getRandomItems};
